omnistream-streaming-platform/requirements.txt
dbt-bigquery
kafka-python
requests
psycopg2-binary
python-dotenv